Move SQLPad req/res decorations to own file
